pyproject.toml
[project]
name = "airport-data-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.25.1",
    "azure-storage-file-datalake>=12.23.0",
    "bs4>=0.0.2",
    "click>=8.3.1",
    "curl-cffi>=0.14.0",
    "google-cloud-storage>=3.8.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
